+++ nodes/decoder.js
@@ -595,6 +595,22 @@ module.exports = {
       deviceId: uint32(bytes, 4),
       updated: bool(bytes, 8)
     }
+  },
+
+  /**
+    * Decodes the response to an activate-keypads request (function code 0xa4).
+    *
+    * @param {buffer}   buffer     64 byte NodeJS buffer
+    * @param {function} translator (optional) function to internationalise the text in a
+    *                              decoded object
+    *
+    * @param {object}   Decoded activate-keypads response object
+    */
+  ActivateKeypads: function (bytes, translator) {
+    return {
+      deviceId: uint32(bytes, 4),
+      updated: bool(bytes, 8)
+    }
   }
 }