+++ nodes/opcodes.js
@@ -20,6 +20,7 @@ module.exports = {
   ClearTimeProfiles: 0x8a,
   SetPCControl: 0xa0,
   SetInterlock: 0xa2,
+  ActivateKeypads: 0xa4,
   ClearTaskList: 0xa6,
   AddTask: 0xa8,
   RefreshTaskList: 0xac,